requirements.txt
tensorflow==2.5
numpy>=1.17.0
matplotlib>=3.1.1
Pillow>=6.1.0
scikit-image>=0.15.0
tf_agents >= 0.9.0
airsim >= 1.6.0
